submit D.cpp to 'D - Decayed Bridges' (abc120) [C++14 (GCC 5.4.1)]

+++ 2019/0303_ABC120/D.cpp
@@ -78,7 +78,7 @@ bool is_same_group(int ia, int ib)
   return i2g[ia] == i2g[ib];
 }
 
-vector<int> ans;
+vector<ll> ans;
 
 int main()
 {
